sqlgpt/requirements.txt
gradio
sqlalchemy 
psycopg2
langchain==0.0.240
llama-index==0.7.11.post1
openai==0.27.8
PyYAML==5.4.1
